requirements.txt
asyncio
pyyaml
colorama
curses
sqlite3
logging
json
time
random
hashlib
os
sys
signal
